requirements.txt
asttokens==3.0.0
attrs==25.1.0
beautifulsoup4==4.13.3
blinker==1.9.0
Bootstrap-Flask==2.5.0
bs4==0.0.2
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.1
click==8.3.1
colorama==0.4.6
comm==0.2.2
contourpy==1.3.1
cycler==0.12.1
debugpy==1.8.12
decorator==5.1.1
dominate==2.9.1
dotenv==0.9.9
executing==2.2.0
fastjsonschema==2.21.1
Flask==3.1.2
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.2
fonttools==4.56.0
greenlet==3.2.4
h11==0.16.0
html5lib==1.1
idna==3.10
ipykernel==6.29.5
ipython==8.32.0
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.6
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter_client==8.6.3
jupyter_core==5.7.2
kiwisolver==1.4.8
lxml==5.3.1
MarkupSafe==3.0.3
matplotlib==3.10.0
matplotlib-inline==0.1.7
narwhals==1.26.0
nbformat==5.10.4
nest-asyncio==1.6.0
numpy==2.2.2
opencv-python==4.11.0.86
outcome==1.3.0.post0
packaging==24.2
pandas==2.2.3
parso==0.8.4
pillow==11.1.0
platformdirs==4.3.6
plotly==6.0.0
prompt_toolkit==3.0.50
psutil==6.1.1
pure_eval==0.2.3
pycparser==2.23
Pygments==2.19.1
pyparsing==3.2.1
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.1
pywin32==308
pyzmq==26.2.1
referencing==0.36.2
requests==2.32.3
rpds-py==0.22.3
seaborn==0.13.2
selenium==4.36.0
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
SQLAlchemy==2.0.44
stack-data==0.6.3
tornado==6.4.2
traitlets==5.14.3
trio==0.31.0
trio-websocket==0.12.2
typing_extensions==4.15.0
tzdata==2025.1
urllib3==2.5.0
visitor==0.1.3
wcwidth==0.2.13
webencodings==0.5.1
websocket-client==1.9.0
Werkzeug==3.1.3
wsproto==1.2.0
WTForms==3.2.1
WTForms-SQLAlchemy==0.4.2
